# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FPT Event Management API",
    "version": "1.0.0",
    "description": "API for FPT Event Management System - Go Lambda"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Send OTP to reset password",
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent to email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessage"
                }
              }
            }
          },
          "400": {
            "description": "Email missing or invalid"
          },
          "404": {
            "description": "Email not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset password with email and OTP",
        "operationId": "resetPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessage"
                }
              }
            }
          },
          "400": {
            "description": "Missing data or weak password"
          },
          "401": {
            "description": "Invalid or expired OTP"
          },
          "404": {
            "description": "Email not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/admin/create-account": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "[Admin Only] Create account for Staff/Organizer/Admin",
        "operationId": "adminCreateAccount",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateAccountRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "400": {
            "description": "Format error or duplicate Email/Phone"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: Admin only"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "[Admin] Update user (fullName, phone, role, status, optional password)",
        "operationId": "adminUpdateUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: Admin only"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "[Admin] Soft delete user (status -> INACTIVE)",
        "operationId": "adminDeleteUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Soft deleted"
          },
          "400": {
            "description": "Missing/invalid id"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: Admin only"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/users/staff-organizer": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "[Admin] Get STAFF and ORGANIZER users (2 lists)",
        "description": "Only ADMIN can call. Returns staffList and organizerList.",
        "operationId": "getStaffOrganizer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffOrganizerResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Admin only)"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ResetPasswordRequest": {
        "type": "object",
        "required": [
          "email",
          "otp",
          "newPassword"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        }
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
...
      "StatusMessage": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "AdminCreateAccountRequest": {
        "type": "object",
        "required": [
          "fullName",
          "phone",
          "email",
          "password",
          "role"
        ],
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "phone": {
            "type": "string",
            "example": "0912345678"
          },
          "email": {
            "type": "string",
            "example": "staff@fpt.edu.vn"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Pass123"
          },
          "role": {
            "type": "string",
            "enum": [
              "STAFF",
              "ORGANIZER",
              "ADMIN"
            ],
            "example": "STAFF"
          }
        }
      },
      "AdminUpdateUserRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "fullName": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "phone": {
            "type": "string",
            "example": "0909123456"
          },
          "role": {
            "type": "string",
            "enum": [
              "STAFF",
              "ORGANIZER",
              "ADMIN"
            ],
            "example": "STAFF"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "abc123"
          }
        }
      },
      "StaffOrganizerUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "STAFF",
              "ORGANIZER"
            ]
          },
          "status": {
            "type": "string"
          },
          "wallet": {
            "type": "number"
          }
        }
      },
      "StaffOrganizerResponse": {
        "type": "object",
        "properties": {
          "staffList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StaffOrganizerUser"
            }
          },
          "organizerList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StaffOrganizerUser"
            }
          }
        }
      },
...
    }
  }
}